+++ CHANGELOG.markdown
@@ -8,7 +8,7 @@
 
 ### Changed
 
-- Upgrade to cimgui 1.66.2 / imgui 1.66. **This is a very big update, so there
+- Upgrade to cimgui 1.66.2+ / imgui 1.66b. **This is a very big update, so there
   are a lot of breaking changes**
 
 ## [0.0.21] - 2018-10-11